docs(toggle): rename file
