+++ packages/angular-test-app/src/preview-examples/toggle-disabled.ts
@@ -10,7 +10,7 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-toggle-custom-disabled',
+  selector: 'app-toggle-disabled',
   template: ` <ix-toggle disabled></ix-toggle> `,
 })
 export class ToggleCustomDisabled {}
