+++ packages/html-test-app/src/preview-examples/toggle-disabled.html
@@ -7,7 +7,7 @@ SPDX-License-Identifier: MIT
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Toggle custom disable example</title>
+    <title>Toggle disabled example</title>
   </head>
   <body class="theme-brand-dark">
     <!-- Preview code -->
